ElfLoader:  - Fixed SELF loader.  - Rewrote ELF32, ELF64 loaders.

+++ trunk/rpcs3/Emu/Memory/Memory.h
@@ -26,19 +26,19 @@ public:
 		safe_delete(Video_GPUdata);
 	}
 
-	u8* GetMem(u64& addr);
+	u8* GetMem(u32& addr);
 
-	void Write8(u64 addr, const u8 data);
-	void Write16(u64 addr, const u16 data);
-	void Write32(u64 addr, const u32 data);
-	void Write64(u64 addr, const u64 data);
-	void Write128(u64 addr, const u128 data);
+	void Write8(u32 addr, const u8 data);
+	void Write16(u32 addr, const u16 data);
+	void Write32(u32 addr, const u32 data);
+	void Write64(u32 addr, const u64 data);
+	void Write128(u32 addr, const u128 data);
 
-	u8 Read8(u64 addr);
-	u16 Read16(u64 addr);
-	u32 Read32(u64 addr);
-	u64 Read64(u64 addr);
-	u128 Read128(u64 addr);
+	u8 Read8(u32 addr);
+	u16 Read16(u32 addr);
+	u32 Read32(u32 addr);
+	u64 Read64(u32 addr);
+	u128 Read128(u32 addr);
 };
 
 extern MemoryBase Memory;
